+++ lib/kitchen/directions/bake_notes/bake_autotitled_notes.rb
@@ -37,10 +37,22 @@ module Kitchen
         note.exercises.each do |exercise|
           exercise.problem.wrap_children('div', class: 'os-problem-container')
 
-          exercise.problem.prepend(child:
+          unless exercise.has_class?('unnumbered')
+            exercise.problem.prepend(child:
+              <<~HTML
+                <span class="os-title-label">#{I18n.t(:"exercises.exercise")} </span>
+                <span class="os-number">#{exercise.count_in(:note)}</span>
+              HTML
+            )
+          end
+
+          next unless exercise.solution
+
+          exercise.solution.wrap_children('div', class: 'os-solution-container')
+
+          exercise.solution.prepend(child:
             <<~HTML
-              <span class="os-title-label">#{I18n.t(:"exercises.exercise")} </span>
-              <span class="os-number">#{exercise.count_in(:note)}</span>
+              <span class="os-title-label">#{I18n.t(:"exercises.solution")}</span>
             HTML
           )
         end