+++ spec/directions/bake_notes/bake_autotitled_notes_spec.rb
@@ -75,17 +75,40 @@ RSpec.describe Kitchen::Directions::BakeAutotitledNotes do
     )
   end
 
-  let(:exercise_within_note) do
+  let(:numbered_exercise_within_note) do
     book_containing(html:
       one_chapter_with_one_page_containing(
         <<~HTML
           <div data-type="note" id="untitlednote" class="foo">
-          <p>this is a note</p>
-          <div data-type="exercise" id="3360">
-            <div data-type="problem" id="504">
-              <ul>
-                <li>What do you need to know to perform this analysis at the very minimum?</li>
-              </ul>
+            <p>this is a note</p>
+            <div data-type="exercise" id="3360">
+              <div data-type="problem" id="504">
+                <ul>
+                  <li>What do you need to know to perform this analysis at the very minimum?</li>
+                </ul>
+              </div>
+            </div>
+          </div>
+        HTML
+      )
+    )
+  end
+
+  let(:unnumbered_exercise_within_note) do
+    book_containing(html:
+      one_chapter_with_one_page_containing(
+        <<~HTML
+          <div data-type="note" id="untitlednote2" class="blah">
+            <p>this is a note</p>
+            <div class="unnumbered" data-type="exercise" id="3361">
+              <div data-type="problem" id="505">
+                <ul>
+                  <li>What do you need to know to perform this analysis at the very minimum?</li>
+                </ul>
+              </div>
+              <div data-type="solution" id="506">
+                <p>Something</p>
+              </div>
             </div>
           </div>
         HTML
@@ -99,12 +122,14 @@ RSpec.describe Kitchen::Directions::BakeAutotitledNotes do
       'notes': {
         'foo': 'Bar',
         'baz': 'Baaa',
+        'blah': 'Blah',
         'project': 'Project',
         'media-2': 'Media',
         'interactive': 'Link to Learning'
       },
       'exercises': {
-        'exercise': 'Exercise'
+        'exercise': 'Exercise',
+        'solution': 'Answer'
       }
     })
   end
@@ -228,31 +253,66 @@ RSpec.describe Kitchen::Directions::BakeAutotitledNotes do
     )
   end
 
-  it 'bakes unclassified exercises within autotitled notes' do
-    described_class.v1(book: exercise_within_note, classes: %w[foo], bake_exercises: true)
-    expect(exercise_within_note.body.pages.first.notes).to match_normalized_html(
-      <<~HTML
-        <div class="foo" data-type="note" id="untitlednote">
-          <h3 class="os-title" data-type="title">
-            <span class="os-title-label">Bar</span>
-          </h3>
-          <div class="os-note-body">
-            <p>this is a note</p>
-            <div data-type="exercise" id="3360">
-              <div data-type="problem" id="504">
-                <span class="os-title-label">Exercise </span>
-                <span class="os-number">1</span>
-                <div class="os-problem-container">
-                  <ul>
-                    <li>What do you need to know to perform this analysis at the very minimum?</li>
-                  </ul>
+  context 'when autotitled notes have numbered exercise within' do
+    it 'bakes' do
+      described_class.v1(book: numbered_exercise_within_note, classes: %w[foo], bake_exercises: true)
+      expect(numbered_exercise_within_note.body.pages.first.notes).to match_normalized_html(
+        <<~HTML
+          <div class="foo" data-type="note" id="untitlednote">
+            <h3 class="os-title" data-type="title">
+              <span class="os-title-label">Bar</span>
+            </h3>
+            <div class="os-note-body">
+              <p>this is a note</p>
+              <div data-type="exercise" id="3360">
+                <div data-type="problem" id="504">
+                  <span class="os-title-label">Exercise </span>
+                  <span class="os-number">1</span>
+                  <div class="os-problem-container">
+                    <ul>
+                      <li>What do you need to know to perform this analysis at the very minimum?</li>
+                    </ul>
+                  </div>
                 </div>
               </div>
             </div>
           </div>
-        </div>
-      HTML
-    )
+        HTML
+      )
+    end
+  end
+
+  context 'when autotitled notes have unnumbered exercise within' do
+    it 'bakes' do
+      described_class.v1(book: unnumbered_exercise_within_note, classes: %w[blah], bake_exercises: true)
+      expect(unnumbered_exercise_within_note.body.pages.first.notes).to match_normalized_html(
+        <<~HTML
+          <div class="blah" data-type="note" id="untitlednote2">
+            <h3 class="os-title" data-type="title">
+              <span class="os-title-label">Blah</span>
+            </h3>
+            <div class="os-note-body">
+              <p>this is a note</p>
+              <div class="unnumbered" data-type="exercise" id="3361">
+                <div data-type="problem" id="505">
+                  <div class="os-problem-container">
+                    <ul>
+                      <li>What do you need to know to perform this analysis at the very minimum?</li>
+                    </ul>
+                  </div>
+                </div>
+                <div data-type="solution" id="506">
+                  <span class="os-title-label">Answer</span>
+                  <div class="os-solution-container">
+                    <p>Something</p>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+        HTML
+      )
+    end
   end
 
   context 'when book does not use grammatical cases' do