+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Add `context_lead_text` to translations (minor)
 * Make `ElementBase#search_with` callable from an `ElementEnumerator` (minor)
 * Support top-titled tables in `BakeUnnumberedTables` (minor)
 * Stop `NoteElement#title` from breaking for empty notes (patch)