+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [12.2.0] - 2021-10-1
+
 * Add `context_lead_text` to translations (minor)
 * Make `ElementBase#search_with` callable from an `ElementEnumerator` (minor)
 * Support top-titled tables in `BakeUnnumberedTables` (minor)